+++ src/components/Navbars/AdminNavbar.jsx
@@ -1,20 +1,4 @@
-/*!
 
-=========================================================
-* Light Bootstrap Dashboard React - v1.3.0
-=========================================================
-
-* Product Page: https://www.creative-tim.com/product/light-bootstrap-dashboard-react
-* Copyright 2019 Creative Tim (https://www.creative-tim.com)
-* Licensed under MIT (https://github.com/creativetimofficial/light-bootstrap-dashboard-react/blob/master/LICENSE.md)
-
-* Coded by Creative Tim
-
-=========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-*/
 import React, { Component } from "react";
 import { Navbar } from "react-bootstrap";
 
@@ -49,12 +33,12 @@ class Header extends Component {
       <Navbar fluid>
         <Navbar.Header>
           <Navbar.Brand>
-            <a href="#pablo">{this.props.brandText}</a>
+            <p>{this.props.brandText}</p>
           </Navbar.Brand>
-          <Navbar.Toggle onClick={this.mobileSidebarToggle} />
+          {/* <Navbar.Toggle onClick={this.mobileSidebarToggle} /> */}
         </Navbar.Header>
         <Navbar.Collapse>
-          <AdminNavbarLinks />
+          {/* <AdminNavbarLinks /> */}
         </Navbar.Collapse>
       </Navbar>
     );