+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [1.3.0] 2019-05-03
 ### Bug fixing
-- Renamed `src/layouts/Dashboard/Dashboard.jsx` to `src/layouts/Admin.jsx`
+- Renamed `src/layouts/Dashboard/Dashboard.jsx` to `src/layouts/layout.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Dashboard.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Icons.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Maps.jsx`